Att Produto/Servico para CRUD Atendimento

+++ src/main/java/br/edu/ufape/poo/barbeariaufape/negocio/basica/Produto.java
@@ -1,5 +1,6 @@
 package br.edu.ufape.poo.barbeariaufape.negocio.basica;
 
+import java.math.BigDecimal;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -11,26 +12,20 @@ public class Produto {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-
     private String nome;
-    private double preco;
+    private BigDecimal preco;
     private String descricao;
     private int quantidade;
-    //ESTÁ CERTO ????
-    //quantidade  retira estoque? 
-    // quantidade e estoque estão interligados, se sai 1 da quantidade retira 1 do estoque, quantidade == estoque
+
     public Produto() {
-        
     }
 
-    public Produto(String nome, double preco,String produto,int quantidade) {
+    public Produto(String nome, BigDecimal preco,String produto,int quantidade) {
         this.nome = nome;
         this.preco = preco;
         this.descricao=produto;
         this.quantidade=quantidade;
     }
-
-    // Getters e setters
 
     public Long getId() {
         return id;
@@ -48,22 +43,26 @@ public class Produto {
         this.nome = nome;
     }
 
-    public double getPreco() {
+    public BigDecimal getPreco() {
         return preco;
     }
 
-    public void setPreco(double preco) {
+    public void setPreco(BigDecimal preco) {
         this.preco = preco;
     }
+
     public String getDescricao() {
         return descricao;
     }
+
     public void setDescricao(String descricao){
         this.descricao = descricao;
     }
+
     public int getQuantidade() {
         return quantidade;
     }
+
     public void setQuantidade(int quantidade){
         this.quantidade = quantidade;
     }
